2024103438吴永波/结课作业/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
tensorboard>=2.0.0
numpy
Pillow
opencv-python
tqdm 